refractored code so it is more modular/component like for the university section. Added some css styling changes to university site.
