+++ src/components/University/DonationTabs/donationTabs.css
@@ -2,6 +2,7 @@
   border: 1px solid #dee2e6;
   margin-top: 2rem;
   border-radius: 0.5rem;
+  margin-left: 25px;
 }
 
 .card-header .nav-link {
