--- src/components/University/University.css
@@ -1,8 +1,0 @@
-/* University NavBar */
-.unviersity-navbar {
-  border-bottom: solid 2px #dee2e6;
-}
-
-.university-navbar .nav-link {
-  color: black;
-}
