+++ src/components/University/UniversityPortfolioDashboard/UniversityPortfolioDashboard.css
@@ -1,6 +1,12 @@
 .dashboard-container {
-  background-color: #ffffff;
+  background-color: #f6fafd;
   min-height: 100vh;
+  margin-left: 100px;
+   /* border-top: 2px solid black; */
+}
+.dashboard-top-row {
+  border-top: 2px #dee2e5;
+  padding-top: 1rem; /* Optional spacing */
 }
 
 .dashboard-wrapper {
@@ -8,6 +14,7 @@
   border: 1px solid #ccc;
   border-radius: 0.5rem;
   padding: 1.5rem;
+  margin-left: 25px;
 }
 
 .dashboard-row {
