+++ src/components/University/UniversityStockPlot/UniversityStockPlot.css
@@ -2,4 +2,12 @@
   display: flex;
   justify-content: center;
   align-items: center;
+    overflow: hidden;
+  width: 100%;
+  max-width: 900px;
+  background-color: white;
+  border-radius: 12px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.05);
+  padding: 1rem;
+  overflow-x: auto;
 }
